requirements.txt
httpx>=0.27.0
websockets>=13.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
pytest>=8.3.0
